+++ crates/miden-testing/src/kernel_tests/tx/test_prologue.rs
@@ -119,7 +119,7 @@ async fn test_transaction_prologue() -> anyhow::Result<()> {
     };
 
     let code = "
-        use miden::tx_kernel_core::prologue
+        use $kernel::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -627,7 +627,7 @@ pub async fn create_account_invalid_seed() -> anyhow::Result<()> {
         .build()?;
 
     let code = "
-      use miden::tx_kernel_core::prologue
+      use $kernel::prologue
 
       begin
           exec.prologue::prepare_transaction
@@ -645,8 +645,8 @@ pub async fn create_account_invalid_seed() -> anyhow::Result<()> {
 async fn test_get_blk_version() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
     let code = "
-    use miden::tx_kernel_core::memory
-    use miden::tx_kernel_core::prologue
+    use $kernel::memory
+    use $kernel::prologue
 
     begin
         exec.prologue::prepare_transaction
@@ -671,8 +671,8 @@ async fn test_get_blk_version() -> anyhow::Result<()> {
 async fn test_get_blk_timestamp() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
     let code = "
-    use miden::tx_kernel_core::memory
-    use miden::tx_kernel_core::prologue
+    use $kernel::memory
+    use $kernel::prologue
 
     begin
         exec.prologue::prepare_transaction